streamlit_app/requirements.txt
streamlit==1.37.1
pandas==2.2.2
psycopg2-binary==2.9.9
altair==5.3.0
python-dotenv==1.0.1
